chore: Implement second interation of settings and organization page
A dynamic theme change functionality has been imple
mented. There is now more extraction of widgets and
generally cleaner code.

+++ lib/ui/widgets/layout_widgets.dart
@@ -8,32 +8,27 @@ class SectionBanner extends StatelessWidget {
   final String title;
   final Color bannerColor;
   final Color titleColor;
-  final int fontSize;
+  final double fontSize;
 
-  const SectionBanner({
-    Key key,
-    @required this.title,
-    this.bannerColor,
-    this.titleColor,
-    this.fontSize
-  }) : super(key: key);
-  
+  const SectionBanner(
+      {Key key,
+      @required this.title,
+      this.bannerColor,
+      this.titleColor,
+      this.fontSize})
+      : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-
     final ConcordThemeData _appTheme = locator<ConcordThemeManager>().theme;
-    
+
     return Container(
-      color: bannerColor ?? _appTheme.bannerColor,
-      padding: EdgeInsets.fromLTRB(20, 8, 36, 8),
-      child: Text(
-        title,
-        style: TextStyle(
-          fontSize: fontSize ?? 34,
-          fontWeight: FontWeight.bold,
-          color: titleColor ?? _appTheme.secondaryColor
-        )
-      )
-    );
+        color: bannerColor ?? _appTheme.bannerColor,
+        padding: EdgeInsets.fromLTRB(20, 8, 36, 8),
+        child: Text(title,
+            style: TextStyle(
+                fontSize: fontSize ?? 34,
+                fontWeight: FontWeight.bold,
+                color: titleColor ?? _appTheme.secondaryColor)));
   }
 }
